Began Cursed, modified CursedConsol and CursedEditor
Started on Cursed which will run the games.  It is now able to display CursedConsole (though it will still need a lot of work).  Changed the all System.Drawing.Color in CursedConsol to System.ConsoleColor, and implemented a method to convert from ConsolColor to Color so it will work with CursedEditor GUI.  CursedEditor.  Also implemented BufferedGraphics for CursedEditor.  CursedEditor no longer displays any sort of graphics.  I have added a core library to be shared by Cursed and CursedEditor.